% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExACvcf.R
\name{ExACvcf}
\alias{ExACvcf}
\title{Annotate VCF file with info from ExAC API bulk queries.}
\usage{
ExACvcf(input_file)
}
\arguments{
\item{input_file}{input VCF file for annotation}
}
\value{
Output is a dataframe containing the following field:
"Chromosome","Position","Reference_Allele", "Alternate_Alleles",
"Type_of_Variation","Depth_of_Sequence_Coverage","Allele_Count_in_Genotypes",
"Variant_Reads_Percentage", "Allele_Frequency","Variant_Consequences",
"Reference_SNP_ID", "Ensembl_Gene_ID", "Ensembl_Transcript_ID"
}
\description{
This library is used for clean up input VCF
and use the Harvard Medical School ExAC Browser API to bulk
query variants.
}
\examples{
library(ExACvcf)
input_file <- system.file("extdata", "challenge_data.vcf", package = "ExACvcf")
output <- ExACvcf(input_file)
write.table(output, "./annotated_challange_data.vcf", sep = "\t", quote=FALSE)
}
\references{
http://exac.hms.harvard.edu
}
\author{
Tiange (Alex) Cui
}
