% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExACvcf.R
\name{parseInfo}
\alias{parseInfo}
\title{Parse "INFO" field from VCF file, seperated by ";" and "="}
\usage{
parseInfo(info_field)
}
\arguments{
\item{info_field}{"INFO" field from VCF file.}
}
\value{
info_vec A vector contains all the info
}
\description{
Parse "INFO" field from VCF file, seperated by ";" and "="
}
\examples{
parseInfo(info_field)
}
